removed validation before save
 * nothing was validated

+++ app/api/chemotion/report_template_api.rb
@@ -10,7 +10,8 @@ module Chemotion
     resource :report_templates do
       desc 'Return list templates'
       get do
-        data = Entities::ReportTemplateEntity.represent(ReportTemplate.all.order(:id), only: [:id, :name, :report_type, :attachment_id ])
+        data = Entities::ReportTemplateEntity.represent(ReportTemplate.all.order(:id),
+                                                        only: %i[id name report_type attachment_id])
         { templates: data }
       end
 
@@ -37,13 +38,13 @@ module Chemotion
         report_template.name = params[:name]
         report_template.report_type = params[:report_type]
         file = params[:file]
-        if file && tempfile = file[:tempfile]
-          attachment_id = report_template.attachment_id
+
+        if file && file[:tempfile]
+
           attachment = Attachment.new(
             bucket: file[:container_id],
             filename: file[:filename],
             key: file[:name],
-            file_path: file[:tempfile],
             created_by: current_user.id,
             created_for: current_user.id,
             content_type: file[:type],
@@ -51,19 +52,12 @@ module Chemotion
           )
 
           begin
-            ActiveRecord::Base.transaction do
-              attachment.save!
-
-              if attachment.valid?
-                report_template.attachment = attachment
-                report_template.save!
-              else
-                raise ActiveRecord::Rollback
-              end
-            end
+            attachment.save!
+            report_template.attachment = attachment
+            report_template.save!
           ensure
-            tempfile.close
-            tempfile.unlink
+            file[:tempfile].close
+            file[:tempfile].unlink
           end
         end
 
@@ -78,40 +72,36 @@ module Chemotion
       post do
         report_template = ReportTemplate.new(
           name: params[:name],
-          report_type: params[:report_type]
+          report_type: params[:report_type],
         )
 
         file = params[:file]
-        if file
-          if tempfile = file[:tempfile]
-            attachment = Attachment.new(
-              bucket: file[:container_id],
-              filename: file[:filename],
-              key: file[:name],
-              file_path: file[:tempfile],
-              created_by: current_user.id,
-              created_for: current_user.id,
-              content_type: file[:type]
-            )
+        if file && tempfile = file[:tempfile]
+          attachment = Attachment.new(
+            bucket: file[:container_id],
+            filename: file[:filename],
+            key: file[:name],
+            file_path: file[:tempfile],
+            created_by: current_user.id,
+            created_for: current_user.id,
+            content_type: file[:type],
+          )
 
-            begin
-              ActiveRecord::Base.transaction do
-                attachment.save!
+          begin
+            ActiveRecord::Base.transaction do
+              attachment.save!
 
-                attachment.attachment_attacher.attach(File.open(file[:tempfile], binmode: true))
+              attachment.attachment_attacher.attach(File.open(file[:tempfile], binmode: true))
 
-                if attachment.valid?
-                  attachment.attachment_attacher.create_derivatives
-                  attachment.save!
-                  report_template.attachment = attachment
-                else
-                  raise ActiveRecord::Rollback
-                end
-              end
-            ensure
-              tempfile.close
-              tempfile.unlink
+              raise ActiveRecord::Rollback unless attachment.valid?
+
+              attachment.attachment_attacher.create_derivatives
+              attachment.save!
+              report_template.attachment = attachment
             end
+          ensure
+            tempfile.close
+            tempfile.unlink
           end
         end
